修复编辑器无法正常更新可行走地格的问题,修复俘虏相关的BUG

+++ Project/Assets/Sango/Scripts/Game/GameSetting.cs
@@ -333,7 +333,7 @@ namespace Sango.Core
         public void ApplyLargeFontSettings()
         {
             PlayerPrefs.SetInt("IsLargeFontEnabled", IsLargeFontEnabled ? 1 : 0);
-            PlayerPrefs.SetFloat("LargeFontScaleFactor", LargeFontScaleFactor);
+            PlayerPrefs.SetInt("LargeFontScaleFactor", LargeFontScaleFactor);
             PlayerPrefs.Save();
         }
